Divide part
