+++ src/main/java/DataCollection/Service/TempDatasService.java
@@ -1,8 +1,9 @@
 package DataCollection.Service;
 
 import DataCollection.domain.TempDatas;
-import DataCollection.repository.DataCollectionRepository;
-import DataCollection.repository.ReturnAllDataRepository;
+import DataCollection.repository.BaseRepository;
+import DataCollection.repository.MongoDBRepository;
+import DataCollection.repository.TempDatasRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -10,23 +11,27 @@ import org.springframework.stereotype.Service;
 public class TempDatasService {
 
     @Autowired
-    DataCollectionRepository dataCollectionRepository;
+    BaseRepository dataCollectionRepository;
 
     @Autowired
-    ReturnAllDataRepository returnAllDataRepository;
+    MongoDBRepository returnAllDataRepository;
+
+    @Autowired
+    TempDatasRepository tempDatasRepository;
+
 
     //임시로 모델링의 적합성에 체크할 데이터 추출
     public TempDatas getTempDatas(long matchId){
-        return dataCollectionRepository.getTempDatas(matchId);
+        return tempDatasRepository.getTempDatas(matchId);
     }
 
     public TempDatas[] getDataList(int DBId){
-        return dataCollectionRepository.getDatasList(DBId);
+        return tempDatasRepository.getDatasList(DBId);
     }
 
     
     //임시 모델링 테스트에 필요한 데이터 mongoDB에 저장
     public void saveAllDatas(int DBId){
-        returnAllDataRepository.saveDatasList(DBId);
+        tempDatasRepository.saveDatasList(DBId);
     }
 }
