+++ src/main/java/DataCollection/domain/LeagueEntryDto.java
@@ -3,8 +3,6 @@ package DataCollection.domain;
 import lombok.Data;
 
 @Data
-    public class LeagueEntryDto {
-    private String tier;
-    private String rank;
+public class LeagueEntryDto {
     private String summonerName;
-    }
+}
