--- src/main/java/DataCollection/domain/NameList.java
@@ -1,12 +1,0 @@
-package DataCollection.domain;
-
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class NameList {
-    private String tier;
-    private String division;
-    private List<String> Names;
-}
