+++ src/main/java/DataCollection/repository/MongoDBRepository.java
@@ -4,20 +4,21 @@ import DataCollection.domain.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Repository;
-
-import java.util.LinkedList;
-import java.util.List;
 
 @Repository
 @Slf4j
-public class ReturnAllDataRepository {
+public class MongoDBRepository {
 
     @Autowired
     private MongoTemplate mongoTemplate;
     @Autowired
-    DataCollectionRepository dataCollectionRepository;
-
+    BaseRepository dataCollectionRepository;
+    @Autowired
+    TempDatasRepository tempDatasRepository;
+/*
    public List returnAllDetails(int DBId) {
         List<MatchDetail> matchDetails = new LinkedList<>();
         log.info("start");
@@ -38,30 +39,38 @@ public class ReturnAllDataRepository {
         }
         log.info("size : {}",matchDetails.size());
         return matchDetails;
+    }*/
+
+    public void saveMatchDetail(MatchDetail matchDetail){
+    MatchDetail saveDetail1 = mongoTemplate.save(matchDetail);
+    log.info("Save: {}", matchDetail.getGameId());
     }
 
-    public void saveDatasList(int DBId) {
-        List<MatchDetail> matchDetails = new LinkedList<>();
-        log.info("start");
-        for (int i = 0; i <= 200; i++) {
-            int temp = DBId + i;
-            log.info("{}", temp);
-            MatchIds matchIds = new MatchIds();
-            if ((matchIds = dataCollectionRepository.findMatchIds(temp)) != null) {
-                for (int j = 0; j < 100; j++) {
-                    try {
-                        MatchDetail matchDetail = dataCollectionRepository.findMatchDetail(matchIds.getMatchIds().get(j));
-                        matchDetails.add(matchDetail);
-                        TempDatas datas = dataCollectionRepository.MatchDetailtoDatas(matchDetail);
-                        mongoTemplate.save(datas);
-                        log.info("i= {} j={} save {}", temp, j, datas);
-                    } catch (NullPointerException ignore) {
-                    }catch (IndexOutOfBoundsException ignore){}
-                }
-            }
-        }
-        log.info("size : {}",matchDetails.size());
+    public MatchDetail findMatchDetail(long matchId){
+        Query query = Query.query(
+                Criteria.where("_id").is(matchId)
+        );
+        MatchDetail matchDetail = mongoTemplate.findOne(query, MatchDetail.class);
+        return matchDetail;
     }
+
+    public MatchIds findMatchIds(int DBId){
+        Query query = Query.query(
+                Criteria.where("_id").is(DBId)
+        );
+        MatchIds matchIds = mongoTemplate.findOne(query, MatchIds.class);
+        return matchIds;
+    }
+
+    public void saveMatchIds(MatchIds matchIds){
+        mongoTemplate.save(matchIds);
+        log.info("Save matchIds{}",matchIds);
+    }
+
+    public void saveTimeLine(TimeLine timeLine){
+        mongoTemplate.save(timeLine);
+    }
+
 
 }
